+++ aws-android-sdk-cognitoidentityprovider/src/main/java/com/amazonaws/mobileconnectors/cognitoidentityprovider/continuations/ChallengeContinuation.java
@@ -26,6 +26,7 @@ import com.amazonaws.mobileconnectors.cognitoidentityprovider.util.CognitoServic
 import com.amazonaws.services.cognitoidentityprovider.model.RespondToAuthChallengeRequest;
 import com.amazonaws.services.cognitoidentityprovider.model.RespondToAuthChallengeResult;
 
+import java.util.Collections;
 import java.util.HashMap;
 import java.util.Map;
 
@@ -52,8 +53,9 @@ public class ChallengeContinuation implements CognitoIdentityProviderContinuatio
     private final CognitoUser user;
     private final String username;
     private final AuthenticationHandler callback;
-    protected Map<String, String> challengeResponses;
     private final boolean runInBackground;
+    protected Map<String, String> challengeResponses;
+    protected Map<String, String> clientMetaData;
 
     /**
      * Constructs a continuation for a challenge to be presented to the user.
@@ -85,6 +87,25 @@ public class ChallengeContinuation implements CognitoIdentityProviderContinuatio
         this.callback = callback;
         this.runInBackground = runInBackground;
         challengeResponses = new HashMap<String, String>();
+    }
+
+    /**
+     * <p>
+     * <code>clientMetadata</code> is a map of custom key-value pairs that you can provide as input for any
+     * custom work flows. Accessor method for <code>clientMetadata</code>.
+     * </p>
+     * @return ClientMetadata
+     */
+    public Map<String, String> getClientMetaData() {
+        return clientMetaData;
+    }
+
+    /**
+     * Mutator for <code>clientMetadata</code>.
+     * @param clientMetaData MetaData to be passed as input to the lambda triggers.
+     */
+    public void setClientMetaData(Map<String, String> clientMetaData) {
+        this.clientMetaData = Collections.unmodifiableMap(clientMetaData);
     }
 
     /**
@@ -140,6 +161,9 @@ public class ChallengeContinuation implements CognitoIdentityProviderContinuatio
         respondToAuthChallengeRequest.setSession(challengeResult.getSession());
         respondToAuthChallengeRequest.setClientId(clientId);
         respondToAuthChallengeRequest.setChallengeResponses(challengeResponses);
+        if (clientMetaData != null) {
+            respondToAuthChallengeRequest.setClientMetadata(clientMetaData);
+        }
         if (runInBackground) {
             new Thread(new Runnable() {
                 @Override