+++ CHANGELOG.md
@@ -10,7 +10,9 @@
       * Added support for [`IdentityProviderType.SignInWithApple`](https://github.com/aws-amplify/aws-sdk-android/pull/1362/files#diff-2949a3798ae1a86ffd84ddf2ee29e229R30).
   - Amazon Comprehend
   - AWS IoT
+  - AWS Key Management Service (KMS)
   - Amazon Kinesis Firehose
+  - Amazon Kinesis Video Streams
   - AWS Lambda
   - Amazon Lex
       * Added support for `sessionId` in `PostContentResult` and `PostTextResult`. See [Amazon Lex SessionId Support](https://github.com/aws-amplify/aws-sdk-android/pull/1366) for details.