+++ CHANGELOG.md
@@ -2,6 +2,11 @@
 
 ## [Release 2.16.5](https://github.com/aws/aws-sdk-android/releases/tag/release_v2.16.5)
 
+### New Features
+
+- **Amazon Kinesis Video Signaling**
+  - Amazon Kinesis Video Signaling Channels supports GetIceServerConfig and SendAlexaOfferToMaster. See [Amazon Kinese Video Signaling Channels Documentation](https://docs.aws.amazon.com/kinesisvideostreams/latest/dg/API_Operations_Amazon_Kinesis_Video_Signaling_Channels.html) for more details
+
 ### Misc. Updates
 
 - Model updates for the following services
@@ -13,6 +18,7 @@
   - AWS Key Management Service (KMS)
   - Amazon Kinesis Firehose
   - Amazon Kinesis Video Streams
+  - Amazon Kinesis Video Signaling
   - AWS Lambda
   - Amazon Lex
       * Added support for `sessionId` in `PostContentResult` and `PostTextResult`. See [Amazon Lex SessionId Support](https://github.com/aws-amplify/aws-sdk-android/pull/1366) for details.